change /api/products api count query and select query
change query dsl to repository
is this good?

I don't know if it's best to paging like this when the condition is on category id, product_image.
Does the product_image table need to hang conditions and import them first and remove them from the app based on category id?